Made some example dialogues
I will focus on Creeper for now, honestly, for this 'tech demo', I will focus all my attention on a mob or two. I hope community will catch up to make the other ones

Or I turned out to be autistic enough to recreate THAT GBNF Monstrosity to turn an AI into script generator. If it comes to that...

...

Goddammit, am I the one cultured here?!